Temporarily pin hdf5 to 1.10.1 for compatibility
As much of the stack is still using `hdf5` version `1.10.1`, pin `vtk`
to that specific version of `hdf5` even though the pinnings now use
`hdf5` version `1.10.2`. Ultimately this should be reverted once the
rest of stack (specifically `libnetcdf` is rebuilt with `hdf5` version
`1.10.2`). This should provide us a way forward rebuilding `vtk` with
newer pinnings in the interim.