Build matrix of MPI variants
Since VTK depends on libnetcdf and hdf5, it will pick only the
`nompi` variant of MPI by default (because this has a higher
build number for both libraries).  However, users may try to run
with libnetcdf versions with MPI support, which may cause problems.

I propose that we build with all 3 MPI variants but an alternative
would be to explicitly only support `nompi`.